Update Lab2_Skeleton_Code.ino
Changing debug print statements so that it is more intuitive to see and understand, specifically the lab 4 ActionRobotSpeed variable. Before, it was printed between ActionServoMove and ActionRobotDrive, which are both one digit prints. The ActionRobotSpeed is a 1 to 3 digit print, so having it mixed in is very confusing. I added spacing, and made it so that the new line is a single "\n" print instead of being tied to a Serial.println() function call. 

Appearance before: 
    Action:012551
Appearance after:
    Action:011 255

In the example above, 255 is the example value of ActionRobotSpeed. In the second example, it is much clearer what is what.